BLUEDOC-148 - Event infrastructure
Also includes a few bug fixes to email and provenance logging.
