+++ lib/tasks/task_logger.rb
@@ -9,7 +9,7 @@ module Deepblue
     # TODO: add flags for turning on and off parts of message
 
     def format_message( _severity, _timestamp, _progname, msg )
-      # "#{timestamp.to_formatted_s(:db)} #{severity} User: #{EmailHelper.user_email} #{msg}\n"
+      # "#{timestamp.to_formatted_s(:db)} #{severity} User: #{Deepblue::EmailHelper.user_email} #{msg}\n"
       "#{msg}\n"
     end
